Solution to day 5 puzzle 1
This had several invalid opcodes during the run with the given
input, but that did not seem to affect to intended operation of
the program; since there was no requirement to do anything with
invalid opcodes, they could potentiall pass unnoticed. The intended
outputs (opcode 4) were all 0, except for the final diagnostic
value, which was accepted as correct.